+++ CHANGES.rst
@@ -36,7 +36,7 @@ Specviz
 Specviz2d
 ^^^^^^^^^
 
-- Implement the Unit Conversion plugin the Specviz2D. [#3253]
+- Implement the Unit Conversion plugin in Specviz2D. [#3253]
 
 API Changes
 -----------